--- Assets/DefaultMaterial.physicsMaterial2D.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a1a78d3e80c4ffe4e8470389fd2bf80b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 6200000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
