--- Assets/PlayerMovement.cs
@@ -1,47 +1,0 @@
-using UnityEngine;
-
-public class PlayerMovement : MonoBehaviour
-{
-    [SerializeField] private float moveSpeed = 5f;
-    [SerializeField] private float jumpForce = 5f;
-    [SerializeField] private LayerMask groundLayer;
-    [SerializeField] private float groundCheckDistance = 0.2f;
-    [SerializeField] private UIManager uiManager;
-    
-    private Rigidbody2D rb;
-    private bool isGrounded;
-    private float horizontalInput;
-    private int coins = 0;
-
-    private void Start()
-    {
-        rb = GetComponent<Rigidbody2D>();
-        uiManager.UpdateCoinDisplay(coins);
-    }
-
-    private void Update()
-    {
-        horizontalInput = Input.GetAxisRaw("Horizontal");
-        isGrounded = Physics2D.Raycast(transform.position, Vector2.down, groundCheckDistance, groundLayer);
-
-        if (Input.GetKeyDown(KeyCode.Space) && isGrounded)
-        {
-            rb.AddForce(Vector2.up * jumpForce, ForceMode2D.Impulse);
-        }
-    }
-
-    private void FixedUpdate()
-    {
-        rb.linearVelocity = new Vector2(horizontalInput * moveSpeed, rb.linearVelocity.y);
-    }
-
-    private void OnTriggerEnter2D(Collider2D other)
-    {
-        if (other.CompareTag("Coin"))
-        {
-            coins++;
-            uiManager.UpdateCoinDisplay(coins);
-            Destroy(other.gameObject);
-        }
-    }
-}
