--- Assets/PlayerMovement.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 935bf081dcf84b24b8053f2c1a38aa9e
