+++ Assets/Scripts/Coin.cs
