+++ Assets/Scripts/Coin.cs.meta
